% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_v22_model.R
\name{wrap_v22_model}
\alias{wrap_v22_model}
\title{Wrap Predict v2.2 script in a web R call returning results in JSON}
\arguments{
\item{age.start}{age at diagnosis between 25 and 85}

\item{screening}{Clinically detected = 0, screen detected = 1}

\item{size}{Tumour size mm between 0 and 500}

\item{grade}{Cancer grade 1,2 or 3}

\item{nodes}{Number of positive nodes 0 to 100. 0.5 is equivalent to 1 with micromets yes in the PREDICT UI}

\item{er}{ER+ = 1, ER- = 0}

\item{her2}{HER2+ = 1, HER2- = 0, missing = 9}

\item{ki67}{KI67+ = 1, KI67- = 0, missing = 9}

\item{generation}{Chemo generation 0, 2 or 3 only}

\item{horm}{Hormone therapy Yes = 1, no = 0}

\item{traz}{Trastuzumab therapy Yes = 1, no = 0}

\item{bis}{Bisphosphonate therapy Yes = 1, no = 0}
}
\value{
JSON encoded results
wrap_v22_model
}
\description{
Takes parameters (that ultimately derive from an OPENCPU POST), and calls the predict model with those parameters. It then prints the results in JSON format so they can be fetched and parsed reasonably easily by a cljs test runner.
}
\examples{
"wrap_v22_model(66,1,12,2,2,1,1,1,3,1,1,1)"
"curl -d'66,1,12,2,2,1,1,1,3,1,1,1' /ocpu/library/v22/R/wrap_v22_model"
}
\seealso{
v22_model
}
\keyword{Predict}
